+++ packages/docs/site/src/routes/compounds/time_region_value/_navigation.js
@@ -1,0 +1,1 @@
+export const hrefBase = 'compounds/time_region_value';