+++ packages/docs/site/src/routes/compounds/time_region_value/index.svelte
@@ -1,0 +1,15 @@
+<script>
+	import Index from '@svizzle/time_region_value/src/routes/index.svelte';
+
+	import {goto as goTo} from '@sapper/app';
+
+	import {groupsStore, yearRangeStore} from './_data/stores';
+	import {hrefBase} from './_navigation';
+</script>
+
+<Index
+	{goTo}
+	{groupsStore}
+	{yearRangeStore}
+	{hrefBase}
+/>