+++ packages/components/time_region_value/CHANGELOG.md
@@ -5,3 +5,4 @@
 - add `[id]/index.svelte`
 - add `[id]/[year].svelte`
 - note that for now `rollup-plugin-svelte` is using `emitCss: false`
+- note that this version supports UK only